Aggiunta classe Timer

+++ src/encoder/Encoder.cpp
@@ -1,7 +1,5 @@
 #include "Encoder.hpp"
 
-#include <math.h>
-
 int Encoder::PPR = 0;
 int Encoder::pos = 0;
 int Encoder::MAX_POS = 0;
@@ -25,10 +23,10 @@ void Encoder::callback(int way) {
     }
 }
 
-double Encoder::get_angle() {
+float Encoder::get_angle() {
     return (pos*2*M_PI)/Encoder::PPR;
 }
 
-double Encoder::get_angle_degrees() {
+float Encoder::get_angle_degrees() {
     return (pos*360.0)/Encoder::PPR;;
 }
